requirements.txt
# ClinSense Week 1 - Clinical Text Intelligence & Entity Recognition
# Core ML & NLP
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.8.0
scipy>=1.11.0

# Text processing
transformers>=4.35.0
torch>=2.0.0
tokenizers>=0.15.0
datasets>=2.14.0
peft>=0.7.0
accelerate>=0.24.0

# NER - scispaCy for baseline entity extraction
scispacy>=0.5.0
spacy>=3.7.0
# NER model (drugs, diseases) - required for Streamlit Cloud deployment
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz

# Kaggle dataset download
kaggle>=1.5.0

# Experiment tracking & model registry
wandb>=0.16.0
mlflow>=2.9.0

# Utils
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
joblib>=1.3.0

# Evaluation
seqeval>=1.2.2  # For NER sequence evaluation

# Week 2 - API & Monitoring
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
evidently>=0.4.0

# Streamlit demo
streamlit>=1.28.0
plotly>=5.18.0
